Crate srml_randomness_collective_flip

Randomness Module

The Randomness Collective Flip module provides a random function that generates low-influence random values based on the block hashes from the previous 81 blocks. Low-influence randomness can be useful when defending against relatively weak adversaries.

Public Functions

See the Module struct for details of publicly available functions.

Usage

Prerequisites

Import the Randomness Collective Flip module and derive your module's configuration trait from the system trait.

Example - Get random seed for the current block

use support::{decl_module, dispatch::Result, traits::Randomness};

pub trait Trait: system::Trait {}

decl_module! {
	pub struct Module<T: Trait> for enum Call where origin: T::Origin {
		pub fn random_module_example(origin) -> Result {
			let _random_seed = <srml_randomness_collective_flip::Module<T>>::random_seed();
			Ok(())
		}
	}
}

Structs

Module

Enums

Call